combine index
